requirements.txt
streamlit==1.35.0
pandas==2.2.2
azure-storage-blob==12.19.1
openai==1.55.3
httpx==0.27.2
azure-cosmos==4.9.0
matplotlib>=3.6.0
requests>=2.28.0
uvicorn